refactor(starknet_class_manager): move metric code to its own module
No changes, only move.
